+++ tests/WriterTest.php
@@ -13,6 +13,9 @@ namespace OrkTest\Csv;
 
 use org\bovigo\vfs\vfsStream;
 
+/**
+ * Test the Writer class.
+ */
 class WriterTest extends \PHPUnit\Framework\TestCase
 {
 
@@ -28,7 +31,7 @@ class WriterTest extends \PHPUnit\Framework\TestCase
      *
      * @return string
      */
-    protected function getTempFile()
+    protected function getTempFile(): string
     {
         return $this->vfs->url() . '/' . debug_backtrace()[1]['function'];
     }
@@ -59,7 +62,7 @@ class WriterTest extends \PHPUnit\Framework\TestCase
             'file' => $this->getTempFile(),
             'header' => false,
         ]);
-        $csv->write([1,2,3,4,5]);
+        $csv->write([1, 2, 3, 4, 5]);
     }
 
     /**
@@ -100,7 +103,7 @@ class WriterTest extends \PHPUnit\Framework\TestCase
             'file' => $this->getTempFile(),
             'header' => false,
         ]);
-        $this->assertSame(20, $csv->write([0,1,2,3,4,5,6,7,8,9]));
+        $this->assertSame(20, $csv->write([0, 1, 2, 3, 4, 5, 6, 7, 8, 9]));
     }
 
     /**
@@ -116,7 +119,7 @@ class WriterTest extends \PHPUnit\Framework\TestCase
             'file' => $this->getTempFile(),
             'header' => false,
         ]);
-        $csv->write([1,2,3,4,5]);
+        $csv->write([1, 2, 3, 4, 5]);
         vfsStream::setQuota(0);
     }
 
@@ -176,15 +179,9 @@ class WriterTest extends \PHPUnit\Framework\TestCase
             'one' => 1,
             'two' => 2,
         ]);
-        $csv->write([
-            'one' => 1,
-        ]);
-        $csv->write([
-            'two' => 2,
-        ]);
-        $csv->write([
-            'three' => 3,
-        ]);
+        $csv->write(['one' => 1]);
+        $csv->write(['two' => 2]);
+        $csv->write(['three' => 3]);
         unset($csv);
         $this->assertEquals('46b70113332cec6212541e14dc8f417c', md5_file($this->getTempFile()));
     }
@@ -200,8 +197,8 @@ class WriterTest extends \PHPUnit\Framework\TestCase
             'file' => $this->getTempFile(),
             'header' => false,
         ]);
-        $csv->write([1,2,3,4,5]);
-        $csv->write([6,7,8,9,10]);
+        $csv->write([1, 2, 3, 4, 5]);
+        $csv->write([6, 7, 8, 9, 10]);
         unset($csv);
         $this->assertSame('66f1d63c002cde9257adc36a7ed58c31', md5_file($this->getTempFile()));
     }